Enemy 클래스 수정 / EnemyManager 라운드 별 생성 수정 / EventManager 수정 / 레벨 추가 / GameManager 추가
Enemy 클래스에 Human 타입 추가 / Enemy 객체가 Weapon을 랜덤으로 생성받고 시작
OnEnemyDied 적 사망 시 이벤트 추가 / 1라운드엔 Human 타입 Enemy 생성
라운드 추가
플레이어 레벨 추가
SpawnManager로 통합

+++ PickPocketRogue/Assets/Script/Weapon.cs
@@ -76,7 +76,7 @@ public class Weapon
     public float GetWeaponDmg() {
         return weaponDmg;
     }
-    public float GetWeaponTypeCode() {
+    public int GetWeaponTypeCode() {
         return weaponTypeCode;
     }
     public string GetWeaponType() {
